improved the performance of the sliding menu's

+++ Test code/fnirsi_1013d_scope/display_lib.h
@@ -28,17 +28,17 @@ struct tagDisplayData
   uint16    *sourcebuffer;         //For copy to screen or slide function the source from where to get the data from
   uint16    *destbuffer;           //For copy from screen the destination where to put the data
   uint16    *linepointer;
-  uint16     xpos;
-  uint16     ypos;
-  uint16     width;
-  uint16     height;
-  uint16     pixelsperline;
+  uint32     xpos;
+  uint32     ypos;
+  uint32     width;
+  uint32     height;
+  uint32     pixelsperline;
 };
 
 //----------------------------------------------------------------------------------------------------------------------------------
 
-void display_set_position(uint16 xpos, uint16 ypos);
-void display_set_dimensions(uint16 width, uint16 height);
+void display_set_position(uint32 xpos, uint32 ypos);
+void display_set_dimensions(uint32 width, uint32 height);
 void display_set_font(PFONTDATA font);
 void display_set_fg_color(uint32 color);
 void display_set_bg_color(uint32 color);
@@ -48,37 +48,37 @@ void display_set_destination_buffer(uint16 *buffer);
 
 //----------------------------------------------------------------------------------------------------------------------------------
 
-void display_draw_line(uint16 xstart, uint16 ystart, uint16 xend, uint16 yend);
-void display_draw_horz_line(uint16 ypos, uint16 xstart, uint16 xend);
-void display_draw_vert_line(uint16 xpos, uint16 ystart, uint16 yend);
-void display_draw_rect(uint16 xpos, uint16 ypos, uint16 width, uint16 height);
-void display_draw_rounded_rect(uint16 xpos, uint16 ypos, uint16 width, uint16 height, uint16 radius);
-void display_draw_arc(uint16 xpos, uint16 ypos, uint16 radius, uint16 startangle, uint16 endangle, uint16 direction);
+void display_draw_line(uint32 xstart, uint32 ystart, uint32 xend, uint32 yend);
+void display_draw_horz_line(uint32 ypos, uint32 xstart, uint32 xend);
+void display_draw_vert_line(uint32 xpos, uint32 ystart, uint32 yend);
+void display_draw_rect(uint32 xpos, uint32 ypos, uint32 width, uint32 height);
+void display_draw_rounded_rect(uint32 xpos, uint32 ypos, uint32 width, uint32 height, uint32 radius);
+void display_draw_arc(uint32 xpos, uint32 ypos, uint32 radius, uint32 startangle, uint32 endangle, uint32 direction);
 
 //----------------------------------------------------------------------------------------------------------------------------------
 
-void display_fill_rect(uint16 xpos, uint16 ypos, uint16 width, uint16 height);
-void display_fill_rounded_rect(uint16 xpos, uint16 ypos, uint16 width, uint16 height, uint16 radius);
+void display_fill_rect(uint32 xpos, uint32 ypos, uint32 width, uint32 height);
+void display_fill_rounded_rect(uint32 xpos, uint32 ypos, uint32 width, uint32 height, uint32 radius);
 
 //----------------------------------------------------------------------------------------------------------------------------------
 
-void display_slide_top_rect_onto_screen(uint16 xpos, uint16 ypos, uint16 width, uint16 height, uint32 speed);
-void display_slide_left_rect_onto_screen(uint16 xpos, uint16 ypos, uint16 width, uint16 height, uint32 speed);
-void display_slide_right_rect_onto_screen(uint16 xpos, uint16 ypos, uint16 width, uint16 height, uint32 speed);
+void display_slide_top_rect_onto_screen(uint32 xpos, uint32 ypos, uint32 width, uint32 height, uint32 speed);
+void display_slide_left_rect_onto_screen(uint32 xpos, uint32 ypos, uint32 width, uint32 height, uint32 speed);
+void display_slide_right_rect_onto_screen(uint32 xpos, uint32 ypos, uint32 width, uint32 height, uint32 speed);
 
 //----------------------------------------------------------------------------------------------------------------------------------
 
-void display_copy_rect_from_screen(uint16 xpos, uint16 ypos, uint16 width, uint16 height);
-void display_copy_rect_to_screen(uint16 xpos, uint16 ypos, uint16 width, uint16 height);
+void display_copy_rect_from_screen(uint32 xpos, uint32 ypos, uint32 width, uint32 height);
+void display_copy_rect_to_screen(uint32 xpos, uint32 ypos, uint32 width, uint32 height);
 
 //----------------------------------------------------------------------------------------------------------------------------------
 
-void display_copy_icon_use_colors(const uint8 *icon, uint16 xpos, uint16 ypos, uint16 width, uint16 height);
+void display_copy_icon_use_colors(const uint8 *icon, uint32 xpos, uint32 ypos, uint32 width, uint32 height);
 
 //----------------------------------------------------------------------------------------------------------------------------------
 
-void display_decimal(uint16 xpos, uint16 ypos, int32 value);
-void display_text(uint16 xpos, uint16 ypos, int8 *text);
+void display_decimal(uint32 xpos, uint32 ypos, int32 value);
+void display_text(uint32 xpos, uint32 ypos, int8 *text);
 
 //----------------------------------------------------------------------------------------------------------------------------------
 //Variable width font handling functions
